Create pageLastNextReview compound type and use it inside guidance document type.
