+++ repository-data/application/src/main/resources/hcm-config/namespaces/hee-cms-platform.cnd
@@ -13,3 +13,6 @@
 [hee:link] > hee:basedocument, hippostd:relaxed, hippotranslation:translated
   orderable
 
+[hee:pageLastNextReview] > hippo:compound, hippostd:relaxed
+  orderable
+
