+++ repository-data/application/src/main/resources/hcm-config/namespaces/hee/pageLastNextReview.yaml
@@ -1,0 +1,68 @@
+definitions:
+  config:
+    /hippo:namespaces/hee/pageLastNextReview:
+      jcr:primaryType: hipposysedit:templatetype
+      jcr:mixinTypes: ['editor:editable', 'mix:referenceable']
+      jcr:uuid: 88b91222-b763-46dd-8dac-8bddf8e85236
+      /hipposysedit:nodetype:
+        jcr:primaryType: hippo:handle
+        jcr:mixinTypes: ['mix:referenceable']
+        jcr:uuid: 0bca1dd1-c3d6-4a5e-b73d-c6ffbdf686df
+        /hipposysedit:nodetype:
+          jcr:primaryType: hipposysedit:nodetype
+          jcr:mixinTypes: ['mix:referenceable', 'hipposysedit:remodel']
+          jcr:uuid: 1bb1035d-ffe3-45bf-8fb5-6174f3b60256
+          hipposysedit:node: true
+          hipposysedit:supertype: ['hippo:compound', 'hippostd:relaxed']
+          hipposysedit:uri: http://www.heecmsplatform.com/hee/nt/1.0
+          /calendardate_:
+            jcr:primaryType: hipposysedit:field
+            hipposysedit:mandatory: false
+            hipposysedit:multiple: false
+            hipposysedit:ordered: false
+            hipposysedit:path: hee:nextReviewed
+            hipposysedit:primary: false
+            hipposysedit:type: CalendarDate
+          /lastReviewed:
+            jcr:primaryType: hipposysedit:field
+            hipposysedit:mandatory: false
+            hipposysedit:multiple: false
+            hipposysedit:ordered: false
+            hipposysedit:path: hee:lastReviewed
+            hipposysedit:primary: false
+            hipposysedit:type: CalendarDate
+      /hipposysedit:prototypes:
+        jcr:primaryType: hipposysedit:prototypeset
+        /hipposysedit:prototype:
+          jcr:primaryType: hee:pageLastNextReview
+          hee:lastReviewed: 0001-01-01T12:00:00Z
+          hee:nextReviewed: 0001-01-01T12:00:00Z
+      /editor:templates:
+        jcr:primaryType: editor:templateset
+        /_default_:
+          jcr:primaryType: frontend:plugincluster
+          frontend:properties: [mode]
+          frontend:references: [wicket.model, model.compareTo, engine, validator.id]
+          frontend:services: [wicket.id, validator.id]
+          /root:
+            jcr:primaryType: frontend:plugin
+            item: ${cluster.id}.field
+            plugin.class: org.hippoecm.frontend.service.render.ListViewPlugin
+          /lastReviewed:
+            jcr:primaryType: frontend:plugin
+            caption: Last Reviewed
+            field: lastReviewed
+            hint: ''
+            plugin.class: org.hippoecm.frontend.editor.plugins.field.PropertyFieldPlugin
+            wicket.id: ${cluster.id}.field
+            /cluster.options:
+              jcr:primaryType: frontend:pluginconfig
+          /nextReviewed:
+            jcr:primaryType: frontend:plugin
+            caption: Next Reviewed
+            field: calendardate_
+            hint: ''
+            plugin.class: org.hippoecm.frontend.editor.plugins.field.PropertyFieldPlugin
+            wicket.id: ${cluster.id}.field
+            /cluster.options:
+              jcr:primaryType: frontend:pluginconfig
